pyproject.toml
[project]
name = "{{template_agent}}"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "contextual-agents-sdk",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langgraph>=0.4.8",
    "uvicorn>=0.34.3",
]
[tool.uv]
override-dependencies = [
]

[tool.uv.sources]
contextual-agents-sdk = { git = "https://github.com/qlik-trial/contextual-agents-sdk", rev = "v1.7.15" }

[dependency-groups]
dev = [
    "ruff>=0.12.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov (>=6.0.0)",
    "yamllint>=1.37.1",
    "pre-commit>=4.3.0",
    "yamlfixer-opt-nc>=0.9.15",
]
